+++ doc/manual.gd
@@ -894,8 +894,8 @@ Source( res ) = N and Range( res ) = N;
 
 #! Let $G$ and $H$ be groups and let $H$ act on $G$ via automorphisms, i.e.
 #! there is a group homomorphism
-#! $$\alpha \colon H \to \operatorname{Aut}(G)$$
-#! such that $g^h = \alpha(h)(g)$ for all $g \in G$ and $h \in H$.
+#! $$\alpha \colon H \to \operatorname{Aut}(G) \colon h \mapsto \alpha_h$$
+#! such that $g^h = \alpha_h(g)$ for all $g \in G$ and $h \in H$.
 #! A **group derivation** $\delta \colon H \to G$ is a map such that
 #! $$\delta(h_1h_2) = \delta(h_1)^{h_2}\delta(h_2).$$
 #! Note that we do not require $G$ to be abelian.
@@ -1034,7 +1034,7 @@ DeclareGlobalFunction( "Image" );
 #! @Arguments der, elm
 DeclareGlobalFunction( "Image" );
 #! @Label of a subgroup under a group derivation
-#! @Arguments der, coll
+#! @Arguments der, sub
 DeclareGlobalFunction( "Image" );
 #! @EndGroup
 
@@ -1131,11 +1131,15 @@ List( ImK );
 
 #! @Chapter Affine actions
 
-#! Let $G$ and $H$ be groups, let $H$ act on $G$ via automorphisms via
+#! Let $G$ and $H$ be groups, let $H$ act on $G$ (via automorphisms) by
 #! $$\alpha \colon H \to \operatorname{Aut}(G) \colon h \mapsto \alpha_h$$
 #! and let $\delta \colon H \to G$ be a group derivation with respect to this
-#! action. Then we can construct a new action, called an **affine action**, by
+#! action. Then we can construct a new action, called the **affine action**
+#! associated to $\delta$, by
 #! $$G \times H \to G \colon g^h = \alpha_h(g) \delta(g).$$
+#! If $K$ is a subgroup of $H$, then the restriction of the affine action of
+#! $H$ on $G$ to $K$ coincides with the affine action of $K$ on $G$ associated
+#! to the restriction of $\delta$ to $K$.
 
 #! <P/>
 
@@ -1151,7 +1155,9 @@ List( ImK );
 #! @Section Creating an affine action
 
 #! @BeginGroup
-#! @Returns the affine action of <A>K</A> via the derivation <A>derv</A>.
+#! @Returns the affine action of <A>K</A> associated to the derivation <A>der</A>.
+#! @Description
+#! The group <A>K</A> must be a subgroup of <C>Source(<A>der</A>)</C>.
 #! @Arguments K, der
 DeclareGlobalFunction( "AffineActionByGroupDerivation" );
 #! @EndGroup
@@ -1167,8 +1173,10 @@ aff := AffineActionByGroupDerivation( H, der );
 
 #! @BeginGroup
 #! @GroupTitle OrbitAffineAction
-#! @Returns the orbit of <A>g</A> under the affine action of <A>K</A> via
-#! <A>der</A>.
+#! @Returns the orbit of <A>g</A> under the affine action of <A>K</A> associated
+#! to <A>der</A>.
+#! @Description
+#! The group <A>K</A> must be a subgroup of <C>Source(<A>der</A>)</C>.
 #! @Arguments K, g, der
 DeclareGlobalFunction( "OrbitAffineAction" );
 #! @EndGroup
@@ -1176,24 +1184,28 @@ DeclareGlobalFunction( "OrbitAffineAction" );
 #! @BeginGroup
 #! @GroupTitle OrbitsAffineAction
 #! @Returns a list containing the orbits under the affine action of <A>K</A>
-#! via <A>der</A> if there are finitely many, or <K>fail</K> if there are
-#! infinitely many.
+#! associated to  <A>der</A> if there are finitely many, or <K>fail</K> if there
+#! are infinitely many.
+#! @Description
+#! The group <A>K</A> must be a subgroup of <C>Source(<A>der</A>)</C>.
 #! @Arguments K, der
 DeclareGlobalFunction( "OrbitsAffineAction" );
 #! @EndGroup
 
 #! @BeginGroup
 #! @GroupTitle NrOrbitsAffineAction
-#! @Returns the number of orbits under the affine action of <A>K</A> via
-#! <A>der</A>.
+#! @Returns the number of orbits under the affine action of <A>K</A> associated
+#! to <A>der</A>.
 #! @Arguments K, der
 DeclareGlobalFunction( "NrOrbitsAffineAction" );
 #! @EndGroup
 
 #! @BeginGroup
 #! @GroupTitle StabiliserAffineAction
-#! @Returns the stabiliser of <A>g</A> under the affine action of <A>K</A> via
-#! <A>der</A>.
+#! @Returns the stabiliser of <A>g</A> under the affine action of <A>K</A> associated
+#! to <A>der</A>.
+#! @Description
+#! The group <A>K</A> must be a subgroup of <C>Source(<A>der</A>)</C>.
 #! @Arguments K, g, der
 DeclareGlobalFunction( "StabiliserAffineAction" );
 #! @Arguments K, g, der
@@ -1203,8 +1215,10 @@ DeclareGlobalFunction( "StabilizerAffineAction" );
 #! @BeginGroup
 #! @GroupTitle RepresentativeAffineAction
 #! @Returns an element of <A>K</A> that maps <A>g1</A> to <A>g2</A> under the
-#! affine action of <A>K</A> via <A>der</A>, or <K>fail</K> if no such element
-#! exists.
+#! affine action of <A>K</A> associated to <A>der</A>, or <K>fail</K> if no such
+#! element exists.
+#! @Description
+#! The group <A>K</A> must be a subgroup of <C>Source(<A>der</A>)</C>.
 #! @Arguments K, g1, g2, der
 DeclareGlobalFunction( "RepresentativeAffineAction" );
 #! @EndGroup
@@ -1225,7 +1239,31 @@ aff( g1, h ) = g2;
 #! true
 #! @EndExample
 
-#! @Section Orbits of affine actions
+#! @Section Operations on orbits of affine actions
+
+#! @BeginGroup
+#! @GroupTitle Representative
+#! @Returns the group element that was used to construct <A>orb</A>.
+#! @Label of an orbit of an affine action
+#! @Arguments orb
+DeclareAttribute( "Representative", IsOrbitAffineActionRep );
+#! @EndGroup
+
+#! @BeginGroup
+#! @GroupTitle ActingDomain
+#! @Returns the group whose affine action <A>orb</A> is an orbit of.
+#! @Label of an orbit of an affine action
+#! @Arguments orb
+DeclareAttribute( "ActingDomain", IsOrbitAffineActionRep );
+#! @EndGroup
+
+#! @BeginGroup
+#! @GroupTitle FunctionAction
+#! @Returns the affine action that <A>orb</A> is an orbit of.
+#! @Label of an orbit of an affine action
+#! @Arguments orb
+DeclareAttribute( "FunctionAction", IsOrbitAffineActionRep );
+#! @EndGroup
 
 #! @BeginGroup
 #! @GroupTitle \in
@@ -1245,6 +1283,15 @@ DeclareAttribute( "Size", IsOrbitAffineActionRep );
 #! @EndGroup
 
 #! @BeginGroup
+#! @GroupTitle StabiliserOfExternalSet
+#! @Returns the stabiliser of <C>Representative(<A>orb</A>)</C> under the
+#! action <C>FunctionAction(<A>orb</A>)</C>.
+#! @Label of an orbit of an affine action
+#! @Arguments orb
+DeclareAttribute( "StabiliserOfExternalSet", IsOrbitAffineActionRep );
+#! @EndGroup
+
+#! @BeginGroup
 #! @GroupTitle List
 #! @Returns a list containing the elements of <A>orb</A>.
 #! @Description
@@ -1254,6 +1301,23 @@ DeclareAttribute( "Size", IsOrbitAffineActionRep );
 #! @Label of an orbit of an affine action
 #! @Arguments orb
 DeclareGlobalFunction( "List" );
+#! @EndGroup
+
+#! @BeginGroup
+#! @GroupTitle Random
+#! @Returns a random element in <A>orb</A>.
+#! @Label in an orbit of an affine action
+#! @Arguments orb
+DeclareOperation( "Random", [ IsOrbitAffineActionRep ] );
+#! @EndGroup
+
+#! @BeginGroup
+#! @GroupTitle \=
+#! @Returns <K>true</K> if <A>orb1</A> is equal to <A>orb2</A>, otherwise
+#! <K>false</K>.
+#! @Label for orbits of an affine action
+#! @Arguments orb1, orb2
+DeclareOperation( "\=", [ IsOrbitAffineActionRep, IsOrbitAffineActionRep ] );
 #! @EndGroup
 
 #! @BeginExample